Move longitude_slicer tests to test_utils
