+++ tests/test_source_bathy.py
@@ -6,7 +6,6 @@ import tempfile
 from pathlib import Path
 import xarray as xr
 from mom6_forge._source_bathy import SourceBathy
-from mom6_forge.utils import longitude_slicer
 
 
 def test_simple_source_bathy_calls(get_rect_topo, synthetic_bathy_file):
@@ -80,76 +79,3 @@ def test_source_bathy_depth_conversion(get_rect_topo, synthetic_bathy_file):
     assert np.any(non_nan_values > 0), "Expected positive depth values for ocean"
 
     assert depth.shape == src.depth.shape
-
-
-def test_longitude_slicer():
-    with pytest.raises(AssertionError):
-        nx, ny, nt = 4, 14, 5
-
-        latitude_extent = (10, 20)
-        longitude_extent = (12, 18)
-
-        dims = ["random_lat", "random_lon", "time"]
-
-        dlambda = (longitude_extent[1] - longitude_extent[0]) / 2
-
-        data = xr.DataArray(
-            np.random.random((ny, nx, nt)),
-            dims=dims,
-            coords={
-                "random_lat": np.linspace(latitude_extent[0], latitude_extent[1], ny),
-                "random_lon": np.array(
-                    [
-                        longitude_extent[0],
-                        longitude_extent[0] + 1.5 * dlambda,
-                        longitude_extent[0] + 2.6 * dlambda,
-                        longitude_extent[1],
-                    ]
-                ),
-                "time": np.linspace(0, 1000, nt),
-            },
-        )
-
-        longitude_slicer(data, longitude_extent, "random_lon")
-
-
-def test_longitude_slicers_regionally():
-    nx, ny = 4, 14
-
-    latitude_extent = (2, 5)
-    longitude_extent = (-90, -70)
-
-    dims = ["random_lat", "random_lon"]
-
-    dlambda = (longitude_extent[1] - longitude_extent[0]) / 2
-
-    data = xr.DataArray(
-        np.random.random((ny, nx)),
-        dims=dims,
-        coords={
-            "random_lat": np.linspace(latitude_extent[0], latitude_extent[1], ny),
-            "random_lon": np.linspace(
-                longitude_extent[0] - 2, longitude_extent[1] + 2, nx
-            ),
-        },
-    )
-
-    # Regular regional
-    data_regular = longitude_slicer(data, longitude_extent, "random_lon")
-    data_east = longitude_slicer(data, (270, 290), "random_lon")
-    assert (data_regular == data_east).all()
-
-    # Seam data
-    longitude_extent = (-5, 5)
-    data = xr.DataArray(
-        np.random.random((ny, nx)),
-        dims=dims,
-        coords={
-            "random_lat": np.linspace(latitude_extent[0], latitude_extent[1], ny),
-            "random_lon": np.linspace(
-                longitude_extent[0] - 2, longitude_extent[1] + 2, nx
-            ),
-        },
-    )
-    data_regular = longitude_slicer(data, longitude_extent, "random_lon")
-    assert len(data_regular.random_lon) > 0
